requirements.txt
pydantic ~= 1.9.0
python-dateutil ~= 2.8.2
python-telegram-bot ~= 13.11
pytz ~= 2021.3
pytz-deprecation-shim ~= 0.1.0.post0
PyYAML ~= 6.0

watchdog ~= 2.1.6
pytest ~= 7.0.1
